feat(sync): add retry mechanism on failed requests (#49)
* feat(sync): retry on 500 error

* feat(sync): refactor retry logic, add to end event

* feat(sync): improve messaging, pull, create, debug

* feat(package): add fs-extra

* feat(sync): improve error messages, default variables, watch_dir

* feat(sync): un-indent command to enable

+++ bin/sync.js
@@ -1,11 +1,11 @@
 #!/usr/bin/env node
 
-const {lstatSync, readdirSync, readFileSync, writeFileSync} = require('fs')
-const {join, parse} = require('path').posix
+const {join, parse, dirname} = require('path').posix
 const readline = require('readline')
 const https = require('https')
 const http = require('http')
 const url = require('url')
+const {lstatSync, readdirSync, readFileSync, ensureDir, exists, readFile, writeFile} = require('fs-extra')
 
 const rl = readline.createInterface({
   input: process.stdin,
@@ -33,6 +33,7 @@ if (process.argv[2] === '--help' || process.argv[2] === '-h') {
   console.log('     PUSH=true', '\t\t\t', 'push files to Files API (compare to `git push --force`)')
   console.log('     DELETE_ALL=true', '\t\t', 'remove all files from Files API. USE WITH CAUTION!!!')
   console.log('     NO_PROMPT=true', '\t\t', 'skip console prompt when making destructive actions. USE WITH CAUTION!!!')
+  console.log('     ENABLE=true', '\t\t', 'enable portal on this workspace')
   console.log()
   console.log('     TYPE=<type>', '\t\t', 'type of files in scanned directories, otherwise use directory structure')
   console.log('     INTERVAL=<seconds>', '\t', 'duration of time in-between scans')
@@ -59,12 +60,14 @@ const LFTIMES = {}
 let {
   DIRECTORY, TYPE, INTERVAL, EMOJI,
   WATCH, DELETE_ALL, PULL, PUSH,
-  KA_RBAC_TOKEN, WORKSPACE, NO_PROMPT
+  KA_RBAC_TOKEN, WORKSPACE, NO_PROMPT,
+  ENABLE, DEBUG, KA_API_URL
 } = process.env
 
-const apiURL = WORKSPACE
-  ? (process.env.KA_API_URL || 'http://127.0.0.1:8001') + `/${WORKSPACE}`
-  : (process.env.KA_API_URL || 'http://127.0.0.1:8001') + '/default'
+WORKSPACE = WORKSPACE || 'default'
+KA_API_URL = KA_API_URL || 'http://127.0.0.1:8001'
+const apiURL = KA_API_URL + `/${WORKSPACE}`
+const workspaceApiURL = KA_API_URL + `/workspaces/${WORKSPACE}`
 
 INTERVAL = parseInt(INTERVAL, 10) || 5
 
@@ -73,7 +76,8 @@ if (DIRECTORY[DIRECTORY.length - 1] !== '/') {
   DIRECTORY = DIRECTORY += '/'
 }
 
-let WATCH_DIR = WATCH === 'true'
+
+let WATCH_DIR = isEnvVarTrue(WATCH)
 let CURL_OUTPUT = false
 if (process.argv[2] === '--curl' || process.argv[2] === '-c') {
   CURL_OUTPUT = true
@@ -87,8 +91,10 @@ function isEnvVarTrue (variable) {
 DELETE_ALL = isEnvVarTrue(DELETE_ALL)
 PUSH = isEnvVarTrue(PUSH)
 PULL = isEnvVarTrue(PULL)
-WATCH = isEnvVarTrue(WATCH)
+WATCH = WATCH_DIR
 EMOJI = isEnvVarTrue(EMOJI)
+ENABLE = isEnvVarTrue(ENABLE)
+DEBUG = isEnvVarTrue(DEBUG)
 
 const PROTOCOLS = {
   'http:': {
@@ -181,6 +187,8 @@ async function read (directory, type) {
 
     // Delete remote file if it exists, recreate it from local file
     return Promise.resolve()
+      .then(() => handle(type, filename, path, deleteFile))
+      .then(res => handleResponse(res, 'delete', isDeleted, type, path))
       .then(() => handle(type, filename, path, createFile))
       .then(res => handleResponse(res, 'create', isCreated, type, path))
       .catch(err => console.log(err))
@@ -189,41 +197,46 @@ async function read (directory, type) {
 
 async function write () {
   let files = await getFiles()
-  return Promise.all(
-    files.data.map(file => {
-      let fileName = file.name
-      let contents = file.contents + '\n'
-      let fileType = file.type + 's/'
-      let extension = file.type === 'spec'
-        ? getSpecExtension(contents)
-        : '.hbs'
 
-      let filePath = (DIRECTORY + fileType + fileName + extension)
+  for (var file of files.data) {
+    let fileName = file.name
+    let contents = file.contents + '\n'
+    let fileType = file.type + 's/'
+    let extension = file.type === 'spec'
+      ? (getSpecExtension(contents) || '.json')
+      : '.hbs'
 
-      if (extension === '' || !extension) {
-        console.log(`Error: Not able to determine extension of ${fileName}`)
-      }
+    let filePath = (DIRECTORY + fileType + fileName + extension)
+    let fileDirectory = dirname(filePath)
+    let shouldWriteFile = true
 
-      try {
-        let localContents = readFileSync(filePath, 'utf8')
-
-        if (localContents !== contents) {
-          console.log(`File Updated: ${filePath}`)
+    let fileExists = await exists(filePath)
+    if (fileExists) {
+      let localContents = await readFile(filePath, 'utf8')
+      if (localContents !== contents) {
+        console.log(`File Updated: ${filePath}`)
+      } else {
+        shouldWriteFile = false
+        if (DEBUG) {
+          console.log(`Contents unchanged. Skipping: ${filePath}`)
         }
-      } catch (e) {
-        console.log(`File Created: ${filePath}`)
       }
+    } else {
+      console.log(`File Created: ${filePath}`)
+    }
 
-      try {
-        writeFileSync(filePath, contents, {flag: 'w'})
-        return Promise.resolve()
-      } catch (e) {
-        console.log(e)
-        console.log(`Pull Failed: unable to write: ${filePath}`)
-        return Promise.resolve()
+    try {
+      if (shouldWriteFile) {
+        await ensureDir(fileDirectory)
+        await writeFile(filePath, contents, { flag: 'w' })
       }
-    })
-  )
+    } catch (e) {
+      console.log(`Pull Failed: unable to write: ${filePath}`)
+      console.log(``)
+      console.log(`\t`, e.message)
+      return false
+    }
+  }
 }
 
 // Helpers
@@ -300,6 +313,16 @@ function handle (type, name, path, action) {
   return action(path, join(pathName, parse(name).name), type, auth)
 }
 
+// General API
+function enablePortal () {
+  const reqUrl = workspaceApiURL
+  return httpRequest(reqUrl, 'PATCH', JSON.stringify({
+    config: {
+      portal: true
+    }
+  }))
+}
+
 // File API
 function getFile (path, nameOrId, type, auth) {
   const reqUrl = `${apiURL}/files?name=${nameOrId}`
@@ -350,8 +373,11 @@ function deleteFile (path, nameOrId, type) {
   return httpRequest(reqUrl, 'DELETE')
 }
 
-function httpRequest (reqUrl, method = 'GET', data = '') {
+function httpRequest (reqUrl, method = 'GET', data = '', tries = 0) {
+  tries++
+
   const options = url.parse(reqUrl)
+
   options.method = method
   options.headers = {
     'Content-Type': 'application/json',
@@ -363,19 +389,33 @@ function httpRequest (reqUrl, method = 'GET', data = '') {
   }
 
   options.agent = PROTOCOLS[options.protocol].agent
-  const adapter = PROTOCOLS[options.protocol].adapter
 
+  const adapter = PROTOCOLS[options.protocol].adapter
   return new Promise((resolve, reject) => {
     const req = adapter.request(options, (res) => {
       let body = ''
       res.on('data', (chunk) => (body += chunk.toString('utf8')))
-      res.on('error', reject)
-      res.on('end', () => {
-        if (res.statusCode >= 200 && res.statusCode <= 299) {
-          resolve({ statusCode: res.statusCode, headers: res.headers, body })
-        } else {
-          reject(`Request failed. status: ${res.statusCode}, body: ${body}`)
+      res.on('error', (err) => {
+        // Internal server error generally means database connection exhaustion
+        // lets retry up to two times then bail out.
+        if (res.statusCode === 500 && tries < 2) {
+          return httpRequest(reqUrl, method, data, tries)
         }
+
+        return reject(err)
+      })
+      res.on('end', () => {
+        // Internal server error generally means database connection exhaustion
+        // lets retry up to two times then bail out.
+        if (res.statusCode === 500 && tries < 2) {
+          return httpRequest(reqUrl, method, data, tries)
+        }
+
+        if (res.statusCode >= 200 && res.statusCode <= 299) {
+          return resolve({ statusCode: res.statusCode, headers: res.headers, body })
+        }
+
+        return reject(`Request failed. status: ${res.statusCode}, body: ${body}`)
       })
     })
 
@@ -429,33 +469,71 @@ function getSpecExtension (content) {
 function promptForPermission(prompt) {
   return new Promise((resolve, reject) => {
     rl.question(prompt, (answer) => {
+      answer = answer.toLowerCase()
+
       if (answer === 'y' || answer ==='yes') {
+        console.log('')
         return resolve(true)
       }
-      console.log('action cancelled')
+
+      console.log('')
+      console.log('Action aborted.')
       return resolve(false)
     })
   })
 }
 
+function areYouSure (message) {
+  return NO_PROMPT || promptForPermission(`⚠️  ${message}\n(y/n) `)
+}
+
+function anErrorHasOccurredTryingTo (message, error, url = apiURL) {
+  if (error === `Request failed. status: 404, body: {"message":"Not Found"}`) {
+    error = [
+      `Portal is disabled on this workspace.`,
+      ``,
+      `Run the following command to enable the portal:`,
+      ``,
+      `KA_API_URL=${KA_API_URL} KA_RBAC_TOKEN=${KA_RBAC_TOKEN} WORKSPACE=${WORKSPACE} ENABLE=true ./bin/sync.js`
+    ].join('\n')
+  }
+  
+  console.log(``)
+  console.log(`🔥 An error occurred while trying to ${message} on ${url}:`)
+  console.log(``)
+  console.log(`\t` + (error.message || error))
+}
+
 async function init () {
+  if (ENABLE) {
+    try {
+      console.log(`Enabling portal on "${WORKSPACE || 'default'}"`)
+      await enablePortal()
+      console.log('✅ Portal enabled.')
+    } catch (e) {
+      anErrorHasOccurredTryingTo('enable the portal', e, workspaceApiURL)
+    }
+    process.exit()
+  }
+
   // Delete all files at start if env flag DELETE_ALL is 'true' (converted to boolean locally)
   if (DELETE_ALL) {
-    let proceed = NO_PROMPT || await promptForPermission(`\n!!!WARNING!!!\n\nYou are about to delete all files from ${apiURL}?\nThis is a destructive action and cannot be reversed.\n\nProceed? (y/n).\n`)
+    let proceed = await areYouSure(`Are you sure you want to delete all files from "${apiURL}"?`)
     if (proceed) {
       try {
-        console.log(`deleting all files from: ${apiURL}`)
+        console.log(`Deleting all files from: ${apiURL}`)
         const {data: files} = await getFiles(TYPE)
         await deleteAllFiles(files)
-      } catch (err) {
-        console.log(`Error deleting files: ${apiURL}`)
+        console.log(`Done.`)
+      } catch (e) {
+        anErrorHasOccurredTryingTo('delete all files', e)
       }
     }
     process.exit()
   }
 
   if (WATCH) {
-    let proceed = NO_PROMPT ||  await promptForPermission(`\n!!!WARNING!!!\n\nThis will watch your templates located in ${DIRECTORY} and push changes to ${apiURL} when a change is detected.\nThis is a destructive action and cannot be reversed.\n\nProceed? (y/n).\n`)
+    let proceed = await areYouSure(`Push changes to "${apiURL}" when a change is detected in "${DIRECTORY}"?`)
     if (proceed) {
       if (TYPE) {
         await read(DIRECTORY, TYPE)
@@ -470,21 +548,31 @@ async function init () {
   }
 
   if (PUSH) {
-    let proceed = NO_PROMPT || await promptForPermission(`\n!!!WARNING!!!\n\nYou are about to push all files located in ${DIRECTORY} to ${apiURL}?\nThis will replace remote files that share the same name and cannot be undone!\n\nProceed? (y/n).\n`)
+    let proceed = await areYouSure(`Push all files to "${apiURL}"?`)
     if (proceed) {
-      console.log(`pushing files to: ${apiURL}`)
-      await read(DIRECTORY)
+      try {
+        console.log(`Pushing files to: ${apiURL}`)
+        await read(DIRECTORY)
+        console.log(`Done.`)
+      } catch (e) {
+        anErrorHasOccurredTryingTo('push files', e)
+      }
     }
     process.exit()
   }
 
   if (PULL) {
-    console.log(`pulling files from: ${apiURL}`)
-    await write()
+    try {
+      console.log(`Pulling files from: ${apiURL}`)
+      await write()
+      console.log(`Done.`)
+    } catch (e) {
+      anErrorHasOccurredTryingTo('pulling files', e)
+    }
     process.exit()
   }
 
-  let proceed = await promptForPermission(`\n!!!WARNING!!!\n\nThis will watch your templates located in ${DIRECTORY} and push changes to ${apiURL} when a change is detected.\nThis is a destructive action and cannot be reversed.\n\nProceed? (y/n).\n`)
+  let proceed = await areYouSure(`Push changes to "${apiURL}" when a change is detected in "${DIRECTORY}"?`)
   if (proceed) {
     if (TYPE) {
       await read(DIRECTORY, TYPE)
